scheduler-sync-pro-main/pyproject.toml
[tool.poetry]
name = "scheduler-sync-pro"
version = "0.1.0"
description = ""
authors = ["Salamat <102606434+skuantaiuly@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.2"}
apscheduler = "^3.10.4"
pytz = "^2024.1"
pymodbus = "^3.6.8"
prettytable = "^3.10.0"
python-snap7 = "^1.4.1"
matplotlib = "^3.9.0"
opcua = "^0.98.13"
pytelegrambotapi = "^4.26.0"
pymongo = "^4.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
